+++ pytest_bazel/__main__.py
@@ -1,3 +1,5 @@
+"""Run the pytest bazel wrapper."""
+
 from pytest_bazel.main import main
 
 main()